refactor: extract context search
